+++ controllers/onlineUsersController.js
@@ -1,10 +1,10 @@
-const webchatModel = require('../models/webchatModel');
+const onlineUsersModel = require('../models/onlineUsersModel');
 
-const addOnlineUser = (payload) => webchatModel.addOnlineUser(payload);
+const addOnlineUser = (payload) => onlineUsersModel.addOnlineUser(payload);
 
-const removeOnlineUser = (nicknameId) => webchatModel.removeOnlineUser(nicknameId);
+const removeOnlineUser = (nicknameId) => onlineUsersModel.removeOnlineUser(nicknameId);
 
-const changeUserNickname = (payload) => webchatModel.changeUserNickname(payload);
+const changeUserNickname = (payload) => onlineUsersModel.changeUserNickname(payload);
 
 module.exports = {
     addOnlineUser,
