+++ controllers/webchatController.js
@@ -2,7 +2,7 @@ const webchatModel = require('../models/webchatModel');
 
 const root = async (req, res) => {
     const messages = await webchatModel.getMessages();
-    res.render('webchat.ejs', { messages, onlineUsers: webchatModel.onlineUsers });
+    res.render('webchat.ejs', { messages });
 };
 
 const saveMessage = async (payload) => {
